separate query, scroll, and search implementations into own files
